reduce duplicate items and improve time format
